Further changes to support sponza model loading; currently has a runtime failure due to mesh indexing issues

+++ Samples/Desktop/D3D12Raytracing/src/D3D12RaytracingSimpleLighting/RaytracingHlslCompat.h
@@ -18,7 +18,7 @@
 using namespace DirectX;
 
 // Shader will use byte encoding to access indices.
-typedef UINT Index;
+typedef UINT16 Index;
 #endif
 
 struct SceneConstantBuffer
